Hide projectiles from spectators and add getSpectators (#86)
* Hide projectiles from spectators and add getSpectators

Added logic to hide projectiles from spectators when a player shoots a bow. Introduced SpectateManager#getSpectators() to retrieve the current spectator players. Also added @NotNull and @Unmodifiable annotations for improved type safety.

* Remove @Unmodifiable annotation from getSpectators

The @Unmodifiable annotation was removed from the getSpectators method, possibly to allow modification of the returned list or to correct an inaccurate annotation.